Move mock data dir
