+++ src/test/common.mock.go
@@ -1,4 +1,4 @@
-package mock
+package test
 
 import "time"
