+++ src/test/contact/contact.mock.go
@@ -1,17 +1,15 @@
-package mock
+package contact
 
 import (
 	"github.com/samithiwat/samithiwat-backend/src/model"
 	"github.com/stretchr/testify/mock"
 )
 
-type ContactMockRepo struct {
+type MockRepo struct {
 	mock.Mock
-	Cont  *model.Contact
-	Conts []*model.Contact
 }
 
-func (r *ContactMockRepo) FindOne(id int, cont *model.Contact) error {
+func (r *MockRepo) FindOne(id int, cont *model.Contact) error {
 	args := r.Called(id, cont)
 	if args.Get(0) != nil {
 		*cont = *args.Get(0).(*model.Contact)
@@ -20,7 +18,7 @@ func (r *ContactMockRepo) FindOne(id int, cont *model.Contact) error {
 	return args.Error(1)
 }
 
-func (r *ContactMockRepo) FindMulti(ids []uint32, conts *[]*model.Contact) error {
+func (r *MockRepo) FindMulti(ids []uint32, conts *[]*model.Contact) error {
 	args := r.Called(ids, conts)
 	if args.Get(0) != nil {
 		*conts = args.Get(0).([]*model.Contact)
@@ -29,7 +27,7 @@ func (r *ContactMockRepo) FindMulti(ids []uint32, conts *[]*model.Contact) error
 	return args.Error(1)
 }
 
-func (r *ContactMockRepo) Create(cont *model.Contact) error {
+func (r *MockRepo) Create(cont *model.Contact) error {
 	args := r.Called(cont)
 	if args.Get(0) != nil {
 		*cont = *args.Get(0).(*model.Contact)
@@ -38,7 +36,7 @@ func (r *ContactMockRepo) Create(cont *model.Contact) error {
 	return args.Error(1)
 }
 
-func (r *ContactMockRepo) Update(id int, cont *model.Contact) error {
+func (r *MockRepo) Update(id int, cont *model.Contact) error {
 	args := r.Called(id, cont)
 	if args.Get(0) != nil {
 		*cont = *args.Get(0).(*model.Contact)
@@ -47,7 +45,7 @@ func (r *ContactMockRepo) Update(id int, cont *model.Contact) error {
 	return args.Error(1)
 }
 
-func (r *ContactMockRepo) Delete(id int, cont *model.Contact) error {
+func (r *MockRepo) Delete(id int, cont *model.Contact) error {
 	args := r.Called(id, cont)
 	if args.Get(0) != nil {
 		*cont = *args.Get(0).(*model.Contact)
